// frontend/src/gen/payment_pb.ts
export enum PaymentMethod {
  /**
   * @generated from enum value: PAYMENT_METHOD_CASH = 0;
   */
  CASH = 0,

  /**
   * @generated from enum value: PAYMENT_METHOD_CARD = 1;
   */
  CARD = 1,

  /**
   * @generated from enum value: PAYMENT_METHOD_TRANSFER = 2;
   */
  TRANSFER = 2,

  /**
   * @generated from enum value: PAYMENT_METHOD_MOMO = 3;
   */
  MOMO = 3,

  /**
   * @generated from enum value: PAYMENT_METHOD_ZALOPAY = 4;
   */
  ZALOPAY = 4,
}